refact(DS-90): extract common of selects in CommonSelect
